+++ src/components/Card/index.tsx
@@ -1,2 +1,2 @@
-export { Card } from "./Card";
+export { Card, CardProps } from "./Card";
 export { CardStyled } from "./Card.styled";