+++ src/components/Element/index.tsx
@@ -1,1 +1,2 @@
 export { Element } from "./Element";
+export { ElementProps, CommonAndHTMLProps } from "./constants";