+++ src/components/Form/index.tsx
@@ -1,32 +1,33 @@
-export { FormWrapper } from "./FormWrapper/FormWrapper";
+export { FormWrapper, FormWrapperProps } from "./FormWrapper/FormWrapper";
 export { FormWrapperStyled } from "./FormWrapper/FormWrapper.styled";
 
-export { FormItem } from "./FormItem/FormItem";
+export { FormItem, FormItemProps } from "./FormItem/FormItem";
 export { FormItemStyled } from "./FormItem/FormItem.styled";
 
-export { FormItemGroup } from "./FormItemGroup/FormItemGroup";
+export { FormItemGroup, FormItemGroupProps } from "./FormItemGroup/FormItemGroup";
 export { FormItemGroupStyled } from "./FormItemGroup/FormItemGroup.styled";
 
-export { InputField } from "./InputField/InputField";
+export { InputField, InputFieldProps } from "./InputField/InputField";
 export { InputFieldStyled } from "./InputField/InputField.styled";
 
-export { TextArea } from "./TextArea/TextArea";
+export { TextArea, TextareaProps } from "./TextArea/TextArea";
 export { TextAreaStyled } from "./TextArea/TextArea.styled";
 
-export { InputLabel } from "./InputLabel/InputLabel";
+export { InputLabel, InputLabelProps } from "./InputLabel/InputLabel";
 
-export { Select } from "./Select/Select";
+export { Select, SelectProps } from "./Select/Select";
 export { SelectStyled } from "./Select/Select.styled";
 
-export { FileUpload } from "./InputField/FileUpload";
+export { FileUpload, FileUploadProps } from "./InputField/FileUpload";
 
 export { RadioGroup } from "./RadioButton/RadioGroup";
+export { RadioButtonProps, RadioGroupProps } from "./RadioButton/constants";
 
 export { RadioButton } from "./RadioButton/RadioButton";
 export { RadioButtonStyled } from "./RadioButton/RadioButton.styled";
 
-export { CheckBox } from "./CheckBox/CheckBox";
+export { CheckBox, CheckboxProps } from "./CheckBox/CheckBox";
 export { CheckBoxStyled } from "./CheckBox/CheckBox.styled";
 
-export { Switch } from "./CheckBox/Switch";
+export { Switch, SwitchProps } from "./CheckBox/Switch";
 export { SwitchStyled } from "./CheckBox/CheckBox.styled";