+++ src/components/Notification/index.tsx
@@ -1,5 +1,5 @@
-export { NotificationsWrapper } from "./NotificationsWrapper/NotificationsWrapper";
+export { NotificationsWrapper, NotificationsWrapperProps } from "./NotificationsWrapper/NotificationsWrapper";
 export { NotificationsWrapperStyled } from "./NotificationsWrapper/NotificationsWrapper.styled";
 
-export { NotificationItem } from "./NotificationItem/NotificationItem";
+export { NotificationItem, NotificationItemProps } from "./NotificationItem/NotificationItem";
 export { NotificationItemStyled } from "./NotificationItem/NotificationItem.styled";