+++ src/components/Portion/index.tsx
@@ -1,2 +1,2 @@
-export { Portion } from "./Portion";
+export { Portion, PortionProps } from "./Portion";
 export { PortionStyled } from "./Portion.styled";