+++ src/components/Table/index.tsx
@@ -1,0 +1,5 @@
+export { Table, TableProps } from "./Table";
+export { TableStyled } from "./Table.styled";
+
+export { TablePagination, TablePaginationProps } from "./TablePagination/TablePagination";
+export { TablePaginationStyled } from "./TablePagination/TablePagination.styled";