+++ src/components/Tooltip/index.tsx
@@ -1,2 +1,2 @@
-export { Tooltip } from "./Tooltip";
+export { Tooltip, TooltipProps } from "./Tooltip";
 export { TooltipStyled } from "./Tooltip.styled";